update awk formula
This is the newest version available of Brian Kernighan's original nawk
awk interpreter. It is of interest to awk programmers testing their
programs on different versions of awk. It is installed as 'nawk' and the
man page as 'nawk.1' and so does not conflict with gawk or any other
awk.

It is also installed in libexec/xpg4bin and libexec/xpg4man as 'awk' and
'awk.1' respectively, as the coreutils formula does.

This is noted in the caveats.

The man page is edited to refer to 'nawk' instead of 'awk' and a couple
of SEE ALSO items are also added. The original man page is installed
as-is into libexec/xpg4man .

The github repo promises to collect patches and track new versions, it
uses the "gentoo" branch for this, which is set as head. If the author
adds more non-trivial patches to this branch and tags it I will switch
to using the tarballs from github instead. Currently there are just a
couple of trivial patches for the makefile.

The homepage and source url is set to Brian's site.

I previously tried to submit this to homebrew main here:

https://github.com/Homebrew/homebrew/pull/43140#event-407127068